refactor(xtask): extract validate comparison facets into a module
The validate checks each re-implemented the same comparison and volatile-
field normalization logic inline. Extract a cohesive comparison module with
single-responsibility differ facets:

- content_diff: destination tree structure, file bytes, symlink targets
- metadata_diff: per-entry mode/mtime/uid/gid/nlink
- VolatileNormalizer + first_line_diff: stdout/stderr equality after masking
  volatile numerics (byte counts, offsets, rates, timings)
- exit_code_diff + classify_failure: exit-status parity and the Skip-vs-Fail
  decision for runs that do not exit cleanly
- WireFrameDiffer / OrderedFrameDiffer: seam for a future wire-frame diff
  (ordered diff implemented and tested; frame capture is left to a later PR)

Route verbosity, stats, progress, itemize, and metadata through the new
facets, and re-export content_diff from support so existing callers keep one
import path. Behaviour is unchanged; each facet carries adversarial unit
tests.
