+++ xtask/src/commands/validate/checks/progress.rs
@@ -12,6 +12,7 @@ use std::collections::HashSet;
 use std::path::Path;
 use std::process::Output;
 
+use crate::commands::validate::comparison;
 use crate::commands::validate::support;
 use crate::commands::validate::transport::{Transport, pull_into};
 use crate::commands::validate::{Check, CheckOutcome, ValidateCtx};
@@ -79,7 +80,7 @@ impl Progress {
             ctx.work,
         ) {
             Ok(out) if out.status.success() => out,
-            other => return skip_or_fail(self.name(), label, "upstream", other),
+            other => return comparison::classify_failure(self.name(), label, "upstream", other),
         };
         let oc = match pull_into(
             transport,
@@ -91,7 +92,7 @@ impl Progress {
             ctx.work,
         ) {
             Ok(out) if out.status.success() => out,
-            other => return skip_or_fail(self.name(), label, "oc", other),
+            other => return comparison::classify_failure(self.name(), label, "oc", other),
         };
 
         // Genuine-result guard: both trees must be fully populated.
@@ -151,27 +152,6 @@ pub fn is_good(pat: &str) -> bool {
         let expected = if i % 2 == 0 { b'N' } else { b'P' };
         b == expected
     })
-}
-
-/// Distinguish a genuine divergence from an unrunnable cell (e.g. ssh refused).
-fn skip_or_fail(
-    check: &'static str,
-    label: &str,
-    who: &str,
-    result: Result<Output, crate::error::TaskError>,
-) -> CheckOutcome {
-    match result {
-        Ok(out) => {
-            let stderr = String::from_utf8_lossy(&out.stderr);
-            let code = out.status.code().unwrap_or(-1);
-            CheckOutcome::fail(
-                check,
-                label,
-                format!("{who} exited {code}: {}", stderr.trim()),
-            )
-        }
-        Err(e) => CheckOutcome::skip(check, label, format!("{who} could not run: {e}")),
-    }
 }
 
 /// Build the progress fixture and return the set of top-level file basenames.
