+++ xtask/src/commands/validate/checks/stats.rs
@@ -10,6 +10,7 @@
 use std::collections::BTreeMap;
 use std::path::Path;
 
+use crate::commands::validate::comparison;
 use crate::commands::validate::support;
 use crate::commands::validate::transport::{Transport, pull_into};
 use crate::commands::validate::{Check, CheckOutcome, ValidateCtx};
@@ -99,7 +100,7 @@ impl Stats {
             ctx.work,
         ) {
             Ok(out) if out.status.success() => out,
-            other => return skip_or_fail(self.name(), label, "upstream", other),
+            other => return comparison::classify_failure(self.name(), label, "upstream", other),
         };
         let oc = match pull_into(
             transport,
@@ -111,7 +112,7 @@ impl Stats {
             ctx.work,
         ) {
             Ok(out) if out.status.success() => out,
-            other => return skip_or_fail(self.name(), label, "oc", other),
+            other => return comparison::classify_failure(self.name(), label, "oc", other),
         };
 
         // Genuine-result guard: both trees must be fully populated.
@@ -179,27 +180,6 @@ fn normalize_value(raw: &str) -> String {
     let trimmed = raw.trim();
     let unitless = trimmed.strip_suffix("bytes").unwrap_or(trimmed).trim_end();
     unitless.chars().filter(|c| *c != ',').collect::<String>()
-}
-
-/// Distinguish a genuine divergence from an unrunnable cell (e.g. ssh refused).
-fn skip_or_fail(
-    check: &'static str,
-    label: &str,
-    who: &str,
-    result: Result<std::process::Output, crate::error::TaskError>,
-) -> CheckOutcome {
-    match result {
-        Ok(out) => {
-            let stderr = String::from_utf8_lossy(&out.stderr);
-            let code = out.status.code().unwrap_or(-1);
-            CheckOutcome::fail(
-                check,
-                label,
-                format!("{who} exited {code}: {}", stderr.trim()),
-            )
-        }
-        Err(e) => CheckOutcome::skip(check, label, format!("{who} could not run: {e}")),
-    }
 }
 
 /// Build the deterministic fixture tree under `src`.
